pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sat-tile-stack"
version = "0.1.0"
description = "Generate daily, multi-band Sentinel-2 tile time-stacks around a point from STAC"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  "numpy",
  "pandas",
  "geopandas",
  "xarray",
  "rioxarray",
  "matplotlib",
  "dask[complete]",
  "pystac-client",
  "planetary-computer",
  "rasterio",
  "stackstac",
  "pyproj",
  "shapely",
  "netcdf4",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "flake8",
  "black",
  "mypy",
]

[tool.setuptools]
packages = ["sattilestack"]
